packages_requirements.txt
numpy==1.13.3
pylint==1.7.2
python-dotenv==0.7.1
websocket-client==0.47.0
requests
